New IntimationList component added
