+++ src/components/ComposePage/CreatePipeline/EmailTagbox.module.css
@@ -3,6 +3,22 @@
     height: 3rem;
     /* border: 0px; */
 }
+ 
+.floatlabelbox {
+    display: flex;
+    flex-wrap: wrap;
+    padding: 1rem;
+    border: 1px solid #D9D9D9;
+    border-radius: 4px;
+    width: 58rem;
+    position: relative;
+    border-radius: 6px;
+    border-width: 1px;
+  }
+
+  .floatlabelbox:hover {
+    border: 1px solid #4096FF;
+  }
 
 .tagbox{
     /* display: flex;
